cleanup and organization of some tests
