+++ tests/e2e/taqueria-cli-e2e-smoke.spec.ts
@@ -1,10 +1,7 @@
 import { prepareEnvironment } from '@gmrchk/cli-testing-library';
-import { ExecException } from 'child_process';
-import { exec as exec1, execSync } from 'child_process';
-import fsPromises from 'fs/promises';
+import { exec as exec1 } from 'child_process';
 import util from 'util';
-import * as contents from './data/help-contents/help-contents';
-import { generateTestProject } from './utils/utils';
+
 const exec = util.promisify(exec1);
 
 describe('Smoke Test E2E Testing for Taqueria CLI,', () => {
@@ -37,19 +34,19 @@ describe('Smoke Test E2E Testing for Taqueria CLI,', () => {
 	test('--help will offer the help menu for a non-initialized project', async () => {
 		const { spawn, cleanup } = await prepareEnvironment();
 		const { waitForText } = await spawn('taq', '--help');
-		await waitForText('taq scaffold');
+		await waitForText('taq init [projectDir]');
 
 		await cleanup();
 	});
 
-	// blocked by https://github.com/ecadlabs/taqueria/issues/1635
-	test.skip('--help will offer help for an initialized project', async () => {
+	test('--help will offer help for an initialized project', async () => {
 		const { spawn, execute, cleanup } = await prepareEnvironment();
 		const { waitForText } = await spawn('taq', 'init test-project --debug');
 		await waitForText("Project taq'ified!");
 
-		const { stdout } = await execute('taq', '--help -p test-project');
-		expect(stdout).toContain('taq [command]');
+		const { stdout } = await execute('taq', '--help', './test-project');
+		console.log(stdout);
+		expect(stdout).toContain('taq <command>');
 
 		await cleanup();
 	});
@@ -101,13 +98,49 @@ describe('Smoke Test E2E Testing for Taqueria CLI,', () => {
 		await cleanup();
 	});
 
-	// TODO - blocked by https://github.com/ecadlabs/taqueria/issues/1692
-	test.skip('install a package that does not exist will  error', async () => {
-		const { execute, cleanup, spawn } = await prepareEnvironment();
+	test('(quickstart scenario) - ligo compile will create an artifact from three files ', async () => {
+		const { execute, spawn, cleanup, writeFile, ls } = await prepareEnvironment();
 		const { waitForText } = await spawn('taq', 'init test-project --debug');
 		await waitForText("Project taq'ified!");
-		const { code } = await execute('taq', 'install acoupleofecadhamburgers -p foobar');
-		expect(code).toBe(1);
+
+		const { stdout, stderr } = await execute('taq', 'install ../taqueria-plugin-ligo', './test-project');
+		expect(stdout).toContain('Plugin installed successfully');
+
+		const mligo_file = await (await exec(`cat e2e/data/counter.mligo`)).stdout;
+		await writeFile('./test-project/contracts/counter.mligo', mligo_file);
+
+		const storage_file = await (await exec(`cat e2e/data/counter.storageList.mligo`)).stdout;
+		await writeFile('./test-project/contracts/counter.storageList.mligo', storage_file);
+
+		const permissions_file = await (await exec(`cat e2e/data/counter.parameterList.mligo`)).stdout;
+		await writeFile('./test-project/contracts/counter.parameterList.mligo', permissions_file);
+
+		const {} = await execute('taq', 'add-contract counter.mligo', './test-project');
+		const {} = await execute('taq', 'add-contract counter.parameterList.mligo', './test-project');
+		const {} = await execute('taq', 'add-contract counter.storageList.mligo', './test-project');
+
+		const contracts_list = await ls('./test-project/contracts');
+		expect(contracts_list).toEqual(
+			expect.arrayContaining(['counter.mligo', 'counter.parameterList.mligo', 'counter.storageList.mligo']),
+		);
+
+		const { stdout: stdout1 } = await execute('taq', 'compile counter.mligo', './test-project');
+
+		expect(stdout1).toMatchInlineSnapshot(`
+              [
+                "┌─────────────────────────────┬───────────────────────────────────────────────┐",
+                "│ Contract                    │ Artifact                                      │",
+                "├─────────────────────────────┼───────────────────────────────────────────────┤",
+                "│ counter.mligo               │ artifacts/counter.tz                          │",
+                "├─────────────────────────────┼───────────────────────────────────────────────┤",
+                "│ counter.storageList.mligo   │ artifacts/counter.default_storage.tz          │",
+                "│                             │ artifacts/counter.storage.another_count.tz    │",
+                "├─────────────────────────────┼───────────────────────────────────────────────┤",
+                "│ counter.parameterList.mligo │ artifacts/counter.parameter.increment_by_3.tz │",
+                "└─────────────────────────────┴───────────────────────────────────────────────┘",
+              ]
+          `);
+
 		await cleanup();
 	});
 });
