+++ tests/e2e/taqueria-plugin-taquito.spec.ts
@@ -6,8 +6,7 @@ import { prepareEnvironment } from '@gmrchk/cli-testing-library';
 import path from 'path';
 
 describe('Taquito Plugin E2E testing for Taqueria CLI', () => {
-	// blocked by https://github.com/ecadlabs/taqueria/issues/1635
-	test.skip('deploy will give contextual help', async () => {
+	test('deploy will give contextual help', async () => {
 		const { execute, spawn, cleanup } = await prepareEnvironment();
 		const { waitForText } = await spawn('taq', 'init test-project');
 		await waitForText("Project taq'ified!");
@@ -27,8 +26,7 @@ describe('Taquito Plugin E2E testing for Taqueria CLI', () => {
 		await cleanup();
 	});
 
-	// blocked by https://github.com/ecadlabs/taqueria/issues/1635
-	test.skip('originate will give contextual help', async () => {
+	test('originate will give contextual help', async () => {
 		const { execute, spawn, cleanup } = await prepareEnvironment();
 		const { waitForText } = await spawn('taq', 'init test-project');
 		await waitForText("Project taq'ified!");
@@ -235,82 +233,6 @@ describe('Taquito Plugin E2E testing for Taqueria CLI', () => {
 			'No accounts were instantiated because they were all instantiated in the target environment already',
 		);
 		expect(stderr).toContain('Note: bob is already instantiated in the current environment, "testing"');
-
-		await cleanup();
-	});
-
-	// FLAKY in cicd, example: https://github.com/ecadlabs/taqueria/actions/runs/3770115326/jobs/6409755728
-	//     Expected: ArrayContaining ["│ Account Alias │ Account Address                      │ Mutez Funded │"]
-	// 	   Received: []
-	test.skip('fund will fund instantiated accounts on a network', async () => {
-		const { execute, spawn, cleanup, writeFile, readFile } = await prepareEnvironment();
-		const { waitForText } = await spawn('taq', 'init test-project');
-		await waitForText("Project taq'ified!");
-		const { stdout } = await execute('taq', 'install ../taqueria-plugin-core', './test-project');
-		expect(stdout).toEqual(expect.arrayContaining(['Plugin installed successfully']));
-		const { stdout: stdout1 } = await execute(
-			'taq',
-			'install ../taqueria-plugin-taquito',
-			'./test-project',
-		);
-		expect(stdout1).toEqual(expect.arrayContaining(['Plugin installed successfully']));
-
-		const test_config_file = await (await exec('cat e2e/data/config-taquito-test-environment-low-tez.json')).stdout;
-		await writeFile('./test-project/.taq/config.json', test_config_file);
-
-		const { stdout: stdout2 } = await execute('taq', 'instantiate-account -e testing', './test-project');
-		expect(stdout2).toContain('Please execute "taq fund" targeting the same environment to fund these accounts');
-
-		const configFile = await readFile(path.join('./test-project', '.taq', 'config.json'));
-		const json = JSON.parse(configFile);
-		expect(json).toBeInstanceOf(Object);
-		expect(json).toHaveProperty('accounts');
-
-		const { stdout: stdout3, stderr } = await execute('taq', 'fund -e testing', './test-project');
-		if (stderr.length > 0) console.error(stderr);
-		expect(stdout3).toEqual(
-			expect.arrayContaining(['│ Account Alias │ Account Address                      │ Mutez Funded │']),
-		);
-
-		await cleanup();
-	});
-
-	test.skip('transfer will send mutez from one instantiated account to another', async () => {
-		// FLAKY - https://github.com/ecadlabs/taqueria/issues/1694
-		const { execute, spawn, cleanup, writeFile } = await prepareEnvironment();
-		const { waitForText } = await spawn('taq', 'init test-project');
-		await waitForText("Project taq'ified!");
-		const { stdout } = await execute('taq', 'install ../taqueria-plugin-core', './test-project');
-		expect(stdout).toEqual(expect.arrayContaining(['Plugin installed successfully']));
-		const { stdout: stdout1 } = await execute(
-			'taq',
-			'install ../taqueria-plugin-taquito',
-			'./test-project',
-		);
-		expect(stdout1).toEqual(expect.arrayContaining(['Plugin installed successfully']));
-
-		const test_config_file = await (await exec('cat e2e/data/config-taquito-test-environment-low-tez.json')).stdout;
-		await writeFile('./test-project/.taq/config.json', test_config_file);
-
-		const { stdout: stdout2 } = await execute('taq', 'instantiate-account -e testing', './test-project');
-		expect(stdout2).toContain('Accounts instantiated: bob, alice, john, jane, joe.');
-
-		const { stdout: stdout3 } = await execute('taq', 'fund -e testing', './test-project');
-		expect(stdout3).toEqual(
-			expect.arrayContaining(['│ Account Alias │ Account Address                      │ Mutez Funded │']),
-		);
-
-		const { stdout: stdout4, stderr } = await execute(
-			'taq',
-			'transfer tz3RobfdmYYQaiF5W343wdSiFhwWF2xUfjEy --mutez 100000 --sender bob -e testing',
-			'./test-project',
-		);
-		if (stderr.length > 0) console.error(stderr);
-		expect(stdout4).toEqual(
-			expect.arrayContaining([
-				'│ N/A            │ tz3RobfdmYYQaiF5W343wdSiFhwWF2xUfjEy │ Unit      │ default    │ 100000         │ https://ghostnet.ecadinfra.com │',
-			]),
-		);
 
 		await cleanup();
 	});
